Create new API util - FetchRebalanceTableJobs
